+++ carp_mobile_sensing/lib/runtime/deployment_controller.dart
@@ -4,7 +4,8 @@
  * Use of this source code is governed by a MIT-style license that can be
  * found in the LICENSE file.
  */
-part of runtime;
+
+part of 'runtime.dart';
 
 /// A [SmartphoneDeploymentController] controls the execution of a [SmartphoneDeployment].
 class SmartphoneDeploymentController extends StudyRuntime<DeviceRegistration> {