+++ carp_mobile_sensing/lib/runtime/device_controller.dart
@@ -4,7 +4,8 @@
  * Use of this source code is governed by a MIT-style license that can be
  * found in the LICENSE file.
  */
-part of runtime;
+
+part of 'runtime.dart';
 
 /// A [DeviceController] handles runtime management of all devices and services
 /// connected to this phone, including the phone itself.