+++ carp_mobile_sensing/lib/runtime/notification/notification_controller.dart
@@ -6,8 +6,20 @@
  */
 part of runtime;
 
-/// A controller of user notifications based on [UserTask]s.
-/// Works closely with the [AppTaskController].
+/// A controller of user notifications allow for creating, scheduling, and
+/// canceling user notifications.
+///
+/// This controller serves two purposes:
+///
+/// 1. It is used by the [AppTaskController] to send notification about
+/// [UserTask]s (which are created based on [AppTask] in the [StudyProtocol]).
+/// This happens automatically, if the app task is configured to
+/// send a notification.
+///
+/// 2. It can be used by the app to create, schedule, and cancel app-specific
+/// notifications. This is done using the [createNotification], [scheduleNotification],
+/// and [scheduleRecurrentNotifications] methods, which
+/// creates an immediate, scheduled, or recurrent notification, respectively.
 abstract class NotificationController {
   /// The upper limit of scheduled notification on iOS.
   static const PENDING_NOTIFICATION_LIMIT = 64;
@@ -44,7 +56,7 @@ abstract class NotificationController {
   /// Create an immediate notification with [id], [title], and [body].
   /// If the [id] is not specified, a random id will be generated.
   ///
-  /// Returns the id of the notification.
+  /// Returns the id of the notification created.
   Future<int> createNotification({
     int? id,
     required String title,
@@ -54,12 +66,33 @@ abstract class NotificationController {
   /// Schedule a notification with [id], [title], and [body] at the [schedule] time.
   /// If the [id] is not specified, a random id will be generated.
   ///
-  /// Returns the id of the notification.
+  /// Returns the id of the notification created.
   Future<int> scheduleNotification({
     int? id,
     required String title,
     String? body,
     required DateTime schedule,
+  });
+
+  /// Schedule recurrent notifications with [id], [title], and [body] at the
+  /// [schedule] time.
+  ///
+  /// Allows for daily, weekly, and monthly recurrence according to the [schedule].
+  ///
+  /// Note that [RecurrentScheduledTrigger.separationCount] and
+  /// [RecurrentScheduledTrigger.end] are **not used**, i.e. days /
+  /// weeks / months cannot be skipped in the scheduled and the notifications
+  /// keeps recurring indefinitely. If you want to stop a recurrent notification
+  /// schedule, use the [cancelNotification] method.
+  ///
+  /// If the [id] is not specified, a random id will be generated.
+  ///
+  /// Returns the id of the notification created.
+  Future<int> scheduleRecurrentNotifications({
+    int? id,
+    required String title,
+    String? body,
+    required RecurrentScheduledTrigger schedule,
   });
 
   /// Cancel (i.e., remove) the notification with [id].
@@ -84,8 +117,14 @@ abstract class NotificationController {
 /// A no-operation notification controller that does nothing.
 class NoOpNotificationController implements NotificationController {
   @override
-  Future<int> createNotification(
-          {int? id, required String title, String? body}) async =>
+  Future<void> initialize() async {}
+
+  @override
+  Future<int> createNotification({
+    int? id,
+    required String title,
+    String? body,
+  }) async =>
       0;
 
   @override
@@ -97,18 +136,24 @@ class NoOpNotificationController implements NotificationController {
       0;
 
   @override
-  Future<void> cancelNotification(int id) async {}
-  @override
-  Future<void> cancelTaskNotification(UserTask task) async {}
+  Future<int> scheduleRecurrentNotifications(
+          {int? id,
+          required String title,
+          String? body,
+          required RecurrentScheduledTrigger schedule}) async =>
+      0;
 
   @override
-  Future<void> initialize() async {}
+  Future<void> cancelNotification(int id) async {}
 
   @override
   Future<void> scheduleTaskNotification(UserTask task) async {}
 
   @override
   Future<void> createTaskNotification(UserTask task) async {}
+
+  @override
+  Future<void> cancelTaskNotification(UserTask task) async {}
 
   @override
   Future<int> get pendingNotificationRequestsCount async => 0;