+++ carp_mobile_sensing/lib/sampling_packages/device/device.dart
@@ -1,5 +1,13 @@
+/*
+ * Copyright 2018 Copenhagen Center for Health Technology (CACHET) at the
+ * Technical University of Denmark (DTU).
+ * Use of this source code is governed by a MIT-style license that can be
+ * found in the LICENSE file.
+ */
+
 /// A library containing a sampling package for collecting information from the
 /// device hardware:
+///
 ///  - device info
 ///  - battery status
 ///  - screen events