+++ carp_mobile_sensing/lib/sampling_packages/sensors/light_probe.dart
@@ -5,7 +5,7 @@
  * found in the LICENSE file.
  */
 
-part of sensors;
+part of 'sensors.dart';
 
 /// The [LightProbe] listens to the phone's light sensor typically located
 /// near the front camera.