+++ carp_mobile_sensing/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.5.0
+
+* Improvements to the Notification Controller
+  * support app-specific methods to create, schedule, and cancel app-specific notifications
+  * asks for permission on Android (Issue [#375](https://github.com/cph-cachet/carp.sensing-flutter/issues/375))
+  * tapping a notification takes the user to the right task in the app (Issue [#376](https://github.com/cph-cachet/carp.sensing-flutter/issues/376))
+
 ## 1.4.7
 
 * improvements to handling connected BLE devices